Seperate input graph from item graph
